requirements.txt
azure-search-documents
langchain
langchain-openai
faiss-cpu
openai
pypdf
python-dotenv
langchain-community
tiktoken
rank_bm25
sentence-transformers
unstructured[pdf]
pdfminer.six
pi-heif
streamlit
